image fallback fix
