+++ src/utils/nftFallback.ts
@@ -1,3 +1,5 @@
+import { SyntheticEvent, useCallback, useEffect, useState } from 'react'
+
 const FALLBACK_IMAGE_BASE_PATH = '/images/collectHeroes'
 const FALLBACK_IMAGE_COUNT = 38
 
@@ -18,8 +20,35 @@ export const getRandomNftFallbackSrc = (exclude?: string): string => {
   return buildFallbackSrc(1)
 }
 
+export const useNftFallbackSource = (
+  src?: string,
+  onError?: (event: SyntheticEvent<Element, Event>) => void,
+) => {
+  const [currentSrc, setCurrentSrc] = useState<string>(() => (src ? src : getRandomNftFallbackSrc()))
+
+  useEffect(() => {
+    if (src) {
+      setCurrentSrc(src)
+      return
+    }
+
+    setCurrentSrc((previous) => getRandomNftFallbackSrc(previous))
+  }, [src])
+
+  const handleError = useCallback(
+    (event: SyntheticEvent<Element, Event>) => {
+      setCurrentSrc((previous) => getRandomNftFallbackSrc(previous))
+      if (onError) {
+        onError(event)
+      }
+    },
+    [onError],
+  )
+
+  return { currentSrc, handleError }
+}
+
 export const nftFallbackConstants = {
   basePath: FALLBACK_IMAGE_BASE_PATH,
   count: FALLBACK_IMAGE_COUNT,
 }
-
